+++ CHANGELOG.md
@@ -3,6 +3,8 @@ MINOR
 
 ## X.Y.Z - changes pending release
 
+### StripeConnect
+* [Fixed] Fixed `AccountOnboardingControllerDelegate.accountOnboardingDidExit(_:)` not being called on the first automatic dismissal after onboarding completion.
 ### PaymentSheet
 * [Changed] Refactored `LinkController` SetupIntent confirmation into a separate `confirmSetupIntent` step (private preview).